CPU instructions implementation plus clock simulation and CPU reset fix
Added CPU instructions implementation, mostly conditionals like BEQ,
BNE, etc.

The CPU_exec function now has a loop that simulates the clock of the
CPU; CPU_reset function also gets a loop, because it caused a BUG in
the CPU_exec function, since CPU_exec only executes a instruction if the clock
is set at zero, before the modification CPU_reset leaved the clock at 8
cycles since that is how long the real CPU takes to reset, but it never
decremented it, so in the first call to CPU_exec the instruction would never
be executed.