All addressing modes implemented
All the memory addressing modes are fully implemented, only left to
implement a few CPU instructions.

Also a new function was introduced, MEM_dump_last_six which as its name
suggest prints to console the last bytes of memory, those are important
since addresses that the cpu uses when an interrupt occurs or when it is
reset are stored there.