6502_functional_test.bin passed
All the logic of the cpu is fully implemented.

The code is full of printf's for debbuging purposes until a proper
system monitor is impemented.

The CPU now checks that an invalid instruction is not being executed, if
that is the case a msg gets printed and the CPU waits for user input;
the same occurs if the CPU catches a loop trap (like the ones that happen in the 6502_functional_test.bin)

In this commit ST_FLAG_MASK is no longer available since we could use only
the shift value

Now MEM_write returns the written value; this is useful in some function
that require the written value to set the CPU flags

CPU shifts Instructions gets modified since in previous commits this
weren't working; a lot of improvement can be done still since code gets
repeated sometimes